+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- `ADD` Administrative areas can be imported
 - `ADD` Published message details can be viewed
 - `ADD` Published messages are shown in the all messages list
 - `ADD` Preview deployment cleanup supports custom branches